Camera2: also check for and request WRITE_EXTERNAL_STORAGE permission
We are asking for READ permissions already, but not for WRITE.
In case we do not get all specified permissions pregranted and the app
asks for them again (eg if installing a development version or user
plays with permissions), the app is not functional and crashes when
taking pictures or trying to delete existing ones.

To fix this, also check for WRITE permissions and if not granted, request
them.

Change-Id: I50b3c44375b5dfa2aa92031a17dc5793dede15eb
Signed-off-by: Alexander Martinz <amartinz@shiftphones.com>